+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Current
+
+## Enhancements
+ - improve resume, add download button
+
 # 0.0.5
 
 ## Features